+++ packages/core/src/animation/AnimationCurve/AnimationColorCurve.ts
@@ -3,13 +3,13 @@ import { StaticInterfaceImplement } from "../../base/StaticInterfaceImplement";
 import { AnimationCurveOwner } from "../internal/AnimationCurveOwner/AnimationCurveOwner";
 import { AnimationCurveReferenceOwner } from "../internal/AnimationCurveOwner/AnimationCurveReferenceOwner";
 import { AnimationCurve } from "./AnimationCurve";
-import { AnimationCurveOwnertType } from "./interfaces/IAnimationCurveStatic";
-import { IAnimationReferenceCurveStatic } from "./interfaces/IAnimationReferenceCurveStatic";
+import { AnimationCurveOwnertType } from "./interfaces/IAnimationCurveCalculator";
+import { IAnimationReferenceCurveCalculator } from "./interfaces/IAnimationReferenceCurveCalculator";
 
 /**
  * Store a collection of Keyframes that can be evaluated over time.
  */
-@StaticInterfaceImplement<IAnimationReferenceCurveStatic<Vector4, Color>>()
+@StaticInterfaceImplement<IAnimationReferenceCurveCalculator<Vector4, Color>>()
 export class AnimationColorCurve extends AnimationCurve<Vector4, Color> {
   /** @internal */
   static _ownerType: AnimationCurveOwnertType = AnimationCurveReferenceOwner;
